+++ src/repo/event.ts
@@ -49,6 +49,17 @@ export class MongoRepository {
                     }
                 });
             }
+
+            if ((<any>params).project.id !== undefined && (<any>params).project.id !== null) {
+                if (typeof (<any>params).project.id.$eq === 'string') {
+                    andConditions.push({
+                        'project.id': {
+                            $exists: true,
+                            $eq: (<any>params).project.id.$eq
+                        }
+                    });
+                }
+            }
         }
 
         // theaterプロパティがあればbranchCodeで検索(互換性維持のため)