+++ src/repo/invoice.ts
@@ -16,7 +16,7 @@ export class MongoRepository {
         this.invoiceModel = connection.model(modelName);
     }
 
-    // tslint:disable-next-line:max-func-body-length
+    // tslint:disable-next-line:cyclomatic-complexity max-func-body-length
     public static CREATE_MONGO_CONDITIONS(params: factory.invoice.ISearchConditions) {
         const andConditions: any[] = [];
 
@@ -46,6 +46,17 @@ export class MongoRepository {
                         $in: params.project.ids
                     }
                 });
+            }
+
+            if ((<any>params).project.id !== undefined && (<any>params).project.id !== null) {
+                if (typeof (<any>params).project.id.$eq === 'string') {
+                    andConditions.push({
+                        'project.id': {
+                            $exists: true,
+                            $eq: (<any>params).project.id.$eq
+                        }
+                    });
+                }
             }
         }